Extract project migration text snippets
Add focused characterization for legacy joint field and accessor migration, then move text snippet builders out of the oversized ProjectMigrationManager without changing migration data.
